+++ src/app/app.js
@@ -9,6 +9,7 @@ angular.module('mwPortal', [
 
   'mwPortal.Start',
   'mwPortal.Hero',
+  'mwPortal.SuperPower',
 
   'mwPortal.Main' // the main module should be at the end so that the configurations are executed on top of the other module configurations
 ]);