+++ src/app/modules/super-power/collections/super-powers.collection.js
@@ -1,4 +1,4 @@
-angular.module('mwPortal.Main')
+angular.module('mwPortal.SuperPower')
   .service('SuperPowersCollection', function (MwPortalCollection, SuperPowerModel) {
     return MwPortalCollection.extend({
       endpoint: '/super-powers',